refactor(db): Split lensed_fetcher.rs into directory module
Same pattern as lensed_auto_commit_fetcher/: struct lifecycle in mod.rs,
DocFetcher trait delegates to fetcher.rs impl methods, migration pipeline
in migration.rs.
